move shareable-notice into `templates` folder
